requirements.txt
numpy==2.2.6
Pillow==12.1.1
matplotlib==3.10.8
open3d==0.19.0
torch==2.10.0+cu128
torchvision==0.25.0+cu128
transformers==5.3.0
